refactor(config): split config.py into core/config/ module
Convert core/config.py into a package with two files:

  core/config/database.py  — DatabaseSettings only (DATABASE_URL)
  core/config/app.py       — AppSettings(DatabaseSettings) + settings singleton
  core/config/__init__.py  — exports from app.py only

The key insight: __init__.py imports only from app.py, NOT from database.py.
alembic/env.py imports directly from core.config.database, bypassing __init__.py
entirely. This means importing DatabaseSettings never triggers app.py execution,
so AppSettings() is never instantiated and API key validation never runs in the
migration job context.

All app code (from core.config import settings/AppSettings) is unchanged —
Python resolves these through __init__.py as before.

Only one import site changes:
  alembic/env.py: from core.config import DatabaseSettings
               -> from core.config.database import DatabaseSettings

No circular references: database.py imports nothing from core.config,
app.py imports from core.config.database (direct file, not __init__.py).
