+++ infra/Pulumi.dev.yaml
@@ -2,4 +2,4 @@ encryptionsalt: v1:sW+JQABkRHo=:v1:7J9KOOOv28snXCef:bsFSX1NP/El1IPwSZCzRTWoV29ZT
 config:
   gcp:project: shared-apps-infrastructure
   gcp:region: europe-west2
-  storyengine-infra:image_tag: fa0fb57
+  storyengine-infra:image_tag: f1be102
